# Type-parameterize Core.Box