--- a/base/boot.jl
+++ b/base/boot.jl
@@ -427,10 +427,12 @@ include(m::Module, fname::String) = ccall(:jl_load_, Any, (Any, Any), m, fname)
 
 eval(m::Module, @nospecialize(e)) = ccall(:jl_toplevel_eval_in, Any, (Any, Any), m, e)
 
-mutable struct Box
-    contents::Any
-    Box(@nospecialize(x)) = new(x)
-    Box() = new()
+mutable struct Box{C}
+    contents::C
+    Box{C}(c) where C = new{C}(c)
+    Box(@nospecialize(c)) = new{Any}(c)
+    Box{C}() where C = new{C}()
+    Box() = new{Any}()
 end
 
 # constructors for built-in types