This edit aims to take a small step toward solving some of the issues in #15276.

There are performance optimizations that do not occur when `Box` accesses are merely `typeasserted`, which instead require the box to be type-parameterized ([example](https://github.com/JuliaLang/julia/issues/15276#issuecomment-1396889311)).

This edit type-parameterizes `Core.Box`, and then future improvements leveraging type annotations and inference can build on top of that.