@giordano 2024-02-19
Why this PR is changing permissions to 28 (out of 31) unrelated files, breaking the build process?