@uniment 2024-02-19
This is a practice run as I get warmed up. My setup got messed up somehow, I'm going to scratch the whole thing.