@KristofferC 2024-02-19
In order for something like this to happen I think it needs to be "proved" that the compiler can take advantage of it and actually generate useful parameterized Boxes in cases where it would otherwise not.