@uniment 2024-02-19
Oh, it appears that in Julia 1.10 the performance penalty that existed for unparameterized `Box` has been eliminated

Before:
```julia
julia> VERSION
v"1.9.4"

julia> @btime (()->begin # parameterized box
           x = Ref(1)
           f() = (x[] = x[] + 1; x[])
           for i=1:1000; f() end
           x[]
       end)()
  1.800 ns (0 allocations: 0 bytes)
1001

julia> @btime (()->begin # typeassert'ed box
           x::Int = 1
           f() = (x = x+1; x)
           for i=1:1000; f() end
           x
       end)()
  4.457 μs (490 allocations: 7.66 KiB)
1001
```
Now:
```julia
julia> VERSION
v"1.10.1"

julia> @btime (()->begin # parameterized box
           x = Ref(1)
           f() = (x[] = x[] + 1; x[])
           for i=1:1000; f() end
           x[]
       end)()
  2.000 ns (0 allocations: 0 bytes)
1001

julia> @btime (()->begin # typeassert'ed box
           x::Int = 1
           f() = (x = x+1; x)
           for i=1:1000; f() end
           x
       end)()
  1.800 ns (0 allocations: 0 bytes)
1001
```