@uniment 2024-02-19
On second look, it appears that even in Julia 1.12 there is still a performance penalty for asserted but unparameterized `Box`:

```julia
julia> VERSION
v"1.12.0-DEV.17"

julia> @btime let a::Int=0; for i=1:1000; a+=i end; ()->a+=1 end;
  5.000 μs (970 allocations: 15.16 KiB)
```

Compare with what is possible by type-parameterizing the box:

```julia
julia> @btime let a=Core.Box{Int}(0); for i=1:1000; a.contents+=i end; ()->a.contents+=1 end;
  4.300 ns (1 allocation: 16 bytes)
```
(notice microseconds vs nanoseconds)